bots/decision_bot/requirements.txt
discord.py>=2.3.2
pandas>=2.0.3
python-dotenv>=1.0.0
requests>=2.31.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
chromadb>=0.4.15
langchain>=0.1.0
transformers>=4.35.0
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0 